requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
async-timeout==5.0.1
attrs==25.1.0
blinker==1.8.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.28.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.24.0
networkx==3.1
numpy==2.2.2
packaging==24.2
pandas==2.0.3
pillow==11.1.0
pkgutil_resolve_name==1.3.10
propcache==0.2.1
protobuf==5.29.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.0
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
timm==1.0.14
toml==0.10.2
torch==2.4.1
torchmetrics==1.6.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
watchdog==4.0.2
yarl==1.18.3
zipp==3.20.2
